One gate instead of two, and a throws that could not
Review finding (PR #313): makeRegisterRequest was marked throws but
cannot throw — dropped (the try at its call sites belongs to
register/unregister and stays); and AsyncGate / TestGate were the same
one-shot gate written twice, now unified as
Tests/OnymIOSTests/Support/OneShotGate.swift — lock-backed so
synchronous test code can open it. The reviewer's stub-churn notes
(setUnregisterAnswer reordering, the redundant-but-consistent public
setter) are acknowledged as note-level and left as-is.
